inserting to list while inserting to tree
i feel like this is cheating. I use the sort() method provided by list in the standard library to get the list ordered.. It could be a waste, because what if the user never needs an ordered list of elements in the tree? It'll be wasting resources while inserting to the tree for no good reason..